pyproject.toml
[project]
name = "interior-decorator-app"
version = "0.1.0"
description = "AI Interior Decorator App"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate==0.27.0",
    "controlnet-aux==0.0.7",
    "diffusers==0.32.2",
    "fastapi==0.109.2",
    "huggingface-hub==0.24.7",
    "numpy==1.26.4",
    "opencv-python==4.8.1.78",
    "pillow==10.2.0",
    "pydantic==2.6.1",
    "python-multipart==0.0.9",
    "torch==2.2.0",
    "transformers==4.37.2",
    "uvicorn==0.27.1",
    "urllib3<2.0.0",
    "streamlit==1.31.0",
    "requests==2.31.0",
    "pydantic-settings==2.1.0",
]

[project.scripts]
api = "api.main:start"
app = "app.ui:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
]
